rename files to stick to one style
