+++ src/bot/bot-factory.ts
@@ -1,9 +1,9 @@
 import { Telegraf } from 'telegraf';
-import { UserService } from '../features/user/services/userService.ts';
-import { BotRouterService } from './botRouterService.ts';
+import { UserService } from '../features/user/services/user.service.ts';
+import { BotRouterService } from './bot-router.service.ts';
 import { config } from '../configs/config.ts';
-import { ConversationStateService } from '../shared/conversation-state/conversationStateService.ts';
-import { MedicationService } from '../features/medication/services/medicationService.ts';
+import { ConversationStateService } from '../shared/conversation-state/conversation-state.service.ts';
+import { MedicationService } from '../features/medication/services/medication.service.ts';
 
 export async function createBot(): Promise<Telegraf> {
   const bot = new Telegraf(config.bot.token);
