+++ src/features/user/repositories/user.repository.ts
@@ -1,5 +1,5 @@
 import type { PrismaClient, User } from '@prisma/client';
-import type { UserInfo } from '../models/UserInfo.ts';
+import type { UserInfo } from '../models/user-info.model.ts';
 import prisma from '../../../shared/database/db.ts';
 
 export class UserRepository {
