+++ src/shared/conversation-state/conversation-state-store.interface.ts
