+++ src/shared/database/db.ts
@@ -3,7 +3,6 @@ import { PrismaClient } from '@prisma/client';
 // Create a singleton instance of PrismaClient
 const prisma = new PrismaClient();
 
-// Export the prisma client for use in other modules
 export default prisma;
 
 // Handle graceful shutdown
